Show installation status when installing from nuclide-installer. Inspired by package-deps from steelbrain.
Summary: @​public
https://github.com/facebook/nuclide/issues/184 filed by @​steelbrain requested that we use his
`package-deps` package to improve the UX when `nuclide-installer` is installing. I took a look
at `package-deps` and filed some issues against it (that @​steelbrain dutifully fixed), but
it still did not seem quite right to me.

Instead, I made some tweaks that are similar to what `package-deps` does, but different in ways
that are most important to us. Most notably, we install packages serially and we limit the
number of packages that we try to install simultaneously based on the number of CPUs the user
has. This was straightforward to add using the `PromisePool` we have in `nuclide-commons`.

We also keep the progress and status text in the notification in sync with what is currently
being installed. Because it can take quite a long time to install all of our Nuclide packages,
it's important to keep the user informed.

Also, I noticed that this package didn't seem to have any Flow errors, so I added the
`"flowCheck": true` to `package.json`, as well. The updates to the `.flowconfig` were for
the transitive deps.

Reviewed By: @jamesgpearce

Differential Revision: D2437648