Create a test utility to copy a fixtures directory to a temp directory.
Summary: @​public
We should not be testing fixtures directories in-place: we should be copying
them to a temp directory instead. This creates a utility to facilitate doing
the right thing.

Note that the implementation is somewhat verbose because `nuclide-test-helpers`
is a dependency of `nuclide-commons`, so we cannot use any of the utilities in
`nuclide-commons` when implementing this helper function.

Reviewed By: @jamesgpearce

Differential Revision: D2460487