Remove ReadOnlyTextEditor as a subclass of TextEditor.
Summary: public
Atom has recently deprecated accessing the `TextEditor` constructor
(which admittedly was never a published part of the Atom 1.0 API):

https://github.com/atom/atom/commit/4d2d4c3

To create a `TextEditor`, you are now supposed to use the new
`atom.workspace.buildTextEditor()` method instead.

They have not removed access to `TextEditor` altogether, but they added some new
checks in the constructor that basically force us to go through
the new `atom.workspace.buildTextEditor()` method:

https://github.com/atom/atom/commit/95fbff03

This is a problem for our `ReadOnlyTextEditor` because it subclasses `TextEditor`
and the way it calls `super()` does not satisfy these checks.

This revision removes the way we create a read-only `TextEditor`. It removes the
subclass and instead overwrites the methods that `ReadOnlyTextEditor` was
overriding on a per-instance basis. I also filed https://github.com/atom/atom/issues/9237
to get us to a less hacky solution in the long-term.

Note that we do a runtime check for the presence of `atom.workspace.buildTextEditor()`
to make sure that we work in both older and newer versions of Atom.

Reviewed By: natthu

Differential Revision: D2564228

fb-gh-sync-id: 874dae28561a9d7ffbc726cb09ab93a463e30257