Remove unnecessary `require('atom').TextEditor` calls.
Summary: public
In a number of cases, we were doing this to appease Flow because
Flow did not always honor `import type`. Because accessing the `TextEditor`
constructor now triggers a deprecation warning in Atom, we remove these
to avoid the warning.

This also removes one instance of:

    import type {TextEditor} from 'atom';

But everywhere else in the codebase, we already treat `TextEditor` like
a globally imported symbol, so it seemed to make more sense to be consistent
for now.

Reviewed By: ssorallen

Differential Revision: D2564284

fb-gh-sync-id: 1846690b63523304b97f8871096da88ed0ca10a5