Replace `instanceof TextEditor` with `isTextEditor`.
Summary: public
Because accessing the `TextEditor` constructor will soon be deprecated,
we can avoid the use of `instanceof TextEditor` by doing duck-typing
checks instead. This introduces an `isTextEditor()` helper function that
does just that.

We should probably work with Atom to introduce an official
`atom.workspace.isTextEditor()` method to complement `atom.workspace.buildTextEditor()`
to eliminate this kludge. Given the fact that JavaScript has a built-in
`Array.isArray()` method, this does not seem like an unreasonable request.

Reviewed By: ssorallen

Differential Revision: D2564369

fb-gh-sync-id: 26732cb95732b1dec6afea1fbca115ce803b8ef5