Shortcut to collapse all visible nodes in file tree
Summary: public

I like my file tree to be relevant to what I'm currently working on, and that
means closing directories that I'm not currently using. Doing this manually is a
pain, so this adds a shortcut key that collapses all visible nodes.

This shortcut, combined with 'reveal in file tree', should make it easy to reset
the state of the file tree and focus on the task at hand.

Reviewed By: ssorallen

Differential Revision: D2614599

fb-gh-sync-id: 7fa081b5fedbd85496e4876b4a84c1a2f92f8614