Ability to publish npm packages only
Summary: public

This adds a `--npm-only` flag to the `publish_packages` script that allows us to
do an npm-only public release.

This is useful if packages outside of our codebase want to depend on a Nuclide package, and want a recent version. We can now provide a recent version without doing a full Nuclide release which would have a wider impact.

Reviewed By: jamesgpearce

Differential Revision: D2645011

fb-gh-sync-id: 36e94997a0b46e92d4cf88d63eee24e5d87b65b0