Improve Flow autocomplete with optional function parameters
Summary: public

Often, optional parameters are not used. In these cases, it is annoying to have
to remove them when using autocomplete. This change makes it so that all
trailing optional parameters are selected when you tab over to the last
non-optional parameter, so that when you start typing you overwrite all of the
remaining parameters.

If you do need the optional ones, you can add them back manually or simply
escape from the selection that tabbing puts you into.

Reviewed By: ssorallen

Differential Revision: D2663163

fb-gh-sync-id: 2190fb4aa6a184da52dbc21c9f4cbc5a9a0c0b51