Fix Flow in busy-signal
Summary: public

In the latest Flow built from source, importing types through symlinks works
properly. Doing that instead of putting busy-signal-interfaces in the
.flowconfig fixes some weird, cryptic Flow errors, and is much nicer anyway.

`flowCheck` still can't be enabled here because this causes some failures in
v0.18.1. But it checks cleanly on the latest Flow, so once v0.19 is released we
should be able to add the flag.

Reviewed By: ssorallen

Differential Revision: D2669610

fb-gh-sync-id: 779231753979d929de0bff854018c9697ee2c11e