Fix OCaml hyperclick
Summary: public

Specifically, hyperclick now calls range.containsPoint. Since this wasn't a
Range, that doesn't work.

Hyperclick should probably construct a new range object whenever it receives a
range so that it works properly with range-compatible arrays. But this is a
minimal fix while we have that discussion. And frankly what was here didn't make
a lot of sense anyway.

Reviewed By: phooji

Differential Revision: D2687466

fb-gh-sync-id: 16001f95416e7ac7eb5ae11e8ad63a515fc3d375