Add warning message if fix does not apply
Summary:
public

This won't catch all cases -- there will be pathological cases where the oldText
check may not catch the fact that the fix is being applied in the wrong spot --
but it will catch most cases where some edits have been made which make the fix
inapplicable.

To really get this right we would either have to disallow applying changes if
the document has been edited (which would be very frustrating), or we would have
to track all the edits and update the fix locations. Long term we will go with
the second option, but for now I think this puts us in a reasonable middle
ground.

Reviewed By: mostafaeweda

Differential Revision: D2767374

fb-gh-sync-id: fb26d19a9cbab07943501f9a5eab0deaa996ddfe