Include the level field in Flow errors
Summary:

public

This wasn't serialized because it wasn't included in the type. Flow didn't catch
it because I'm pretty sure we get `any` out of the service framework, and we
aren't good about casting that to a real type.

This did not appear during local development which is why it wasn't caught
sooner.

Reviewed By: matthewwithanm

Differential Revision: D2775738

fb-gh-sync-id: 725c50372c56bb851cadd1d68348788831d81354