Improve validation for autofix application
Summary:
public

This improves the following situation:

```
somecode;

othercode  // <-- missing semicolon (empty oldText!)
```

Hit enter after somecode, which pushes the empty line into the line where
othercode was. Then hit fix. Before, it would insert a semicolon at the end of
the blank line. Now, it applies nothing and adds a popup.

Unfortunately if the line above is as long or longer a rogue semicolon will
still appear. That's not great.

In the near future I will want to track edits of the source files so that I can
intelligently apply fixes like this without relying on guesswork to tell whether
the patch is still valid. Atom's Markers are probably a good choice.

Reviewed By: mostafaeweda

Differential Revision: D2773748

fb-gh-sync-id: b7a03a3a7418fae619404777e13f494f3e9793f6