Remove extraneous bins leftover from UP merge
Summary:
These bins were merged into the root `package.json` as part of the automated scripts that generated UP. It doesn't make sense to expose these except for `nuclide-start-server`. Especially since the way we'll suggest to install the server is via `npm -g install nuclide` (we'd be polluting the user's `/usr/local/bin` with these).

public

Reviewed By: peterhal

Differential Revision: D2798828

fb-gh-sync-id: 26a54006ead11678f731e99988fdaa2c0ccef996