Move around nuclide-debugger-node VendorLibs
Summary:
This diff fixes an issue an with Atom 1.3.0+, where it would incorrectly identify `v8-debug` and `v8-profiler` as incompatible native modules. This is because it search for `build/MODULE_NAME/**/*.node` in `node_modules`.

As a fix, `v8-debug` and `v8-profiler` were moved to `pkg/nuclide/debugger/node/VendorLib` instead of being pseudo modules in `node_modules`. `node-inspector`'s `require.resolve` calls with `v8-debug` and `v8-profiler` were re-written to be relative instead. `node-inspector` was also moved into a subdirectory of `VendorLib` so that it matches the hierarchy that it was in the official release.

This will allow me to re-enable `nuclide-debugger-atom` in the OSS version (disabled in v0.109.0-alpha3).

public

Reviewed By: ssorallen

Differential Revision: D2820881

fb-gh-sync-id: 8a8d5ab68e69811f7b7983d6fa4464ee4636ee48