Create MarkerTracker so that Markers can track autofix regions
Summary:
public

This allows autofixes to be tracked precisely. Instead of blindly trying to
apply the fix to the location we originally received, we can use Atom's Markers
to track the correct location in the presence of unsaved changes. The
MarkerTracker facilitates the creation and cleanup of Markers to this end.

Reviewed By: ssorallen

Differential Revision: D2841624

fb-gh-sync-id: 7070bbc901ca8e0f049f45ed9d344aa5c2398f8a