Improve mouse UX and query efficiency
Summary:
public

{F59924090}

This makes type-hints work the way we want them to.
Removes a bunch of handrolled timeout logic in favor of clean debounce-based solution. Tool tips don't time out automatically anymore, which is the UX we want for an IDE.
In the screencast above, I logged the queries to the type hint provider, so you can see that it works pretty efficiently and does not issue unnecessary, duplicate queries.
The new logic also ensures that hovering over type hints prevents other type hints from popping up, which is especially important for interactive type-hints.

Reviewed By: mostafaeweda

Differential Revision: D2844008

fb-gh-sync-id: b2246ac00c31abcbe174cf66ce0a7d6f4fa485c7