Add sample package demonstrating gadgets usage
Summary:
public

I started building an unrelated package on top of gadgets, and found that I had to do a bit of copying & chopping to figure out the smallest possible gadget. I figured the result might be worth sharing as a building block for others, and double as documentation.

matthewwithanm, I'd love to hear your thoughts on whether this is a good boilerplate version for people to build on.

{F59932255}

Reviewed By: matthewwithanm

Differential Revision: D2849280

fb-gh-sync-id: 9e11492d7af9f34af8e0c48986b063b881759ceb