Add --nodates option to "hg diff" invocation
Summary:
I am running hg version 3.6.3 on Mac OS X 10.11.3 and trying to use it with nuclide 0.115.1 on Atom 1.5.0-beta3. The problem I am seeing is that the git diff package doesn't show gutter annotations when changes to files are made (after saving). Also the status bar isn't showing the +n -n information for changed files.

After some debugging, it appears that the problem is due to the hg command including a date stamp in the header of diff messages.IIUC, the nuclide hg repository support runs the following command: (with HGPLAIN=1 in the environment) "hg diff --unified 0 --noprefix". The problem appears to be that this command produces a header like this:

 diff -r b9e495e9a0d4 src/org/bitbucket/inkytonik/moama/tests/SyntaxAnalyserTests.scala
 --- src/org/bitbucket/inkytonik/moama/tests/SyntaxAnalyserTests.scala	Thu Feb 04 09:14:17 2016 +1100
+++ src/org/bitbucket/inkytonik/moama/tests/SyntaxAnalyserTests.scala	Thu Feb 04 16:48:10 2016 +110

The hg repository code grabs the filename and date as one string
Closes https://github.com/facebook/nuclide/pull/355

Reviewed By: mareksapota-fb

Differential Revision: D2905060

Pulled By: ssorallen

fb-gh-sync-id: e8ca1723d850aacf14d45bc3db9000f864e52856