Mitigate Atom >1.4.0 render / clipping bug
Summary:
public
heisenbug

Definitely a strong contender for the Elusive Bug of the Year Award.
Per the discussion on t9137213, there are discussions of related symptoms all over other Atom packages:

* https://github.com/atom/atom/issues/9892
* https://github.com/atom/one-dark-ui/issues/109
* https://github.com/atom/one-dark-ui/pull/112
* https://github.com/atom/settings-view/issues/703

There's no clear cause, but it seems to be related to `transform` ie. if something gets off-loaded to the GPU. Oddly enough, `quick-open` does not use transform effects. It does however use the (fairly obscure) `-webkit-mask` to blend the tail end of long search results.

First, to get a better chance of reproducing this issue, I did the following:
- use Atom >1.4.0, specifically v1.4.3
- Use the "Atom Dark" UI theme
- queried quick-open separately for all letters a..z and observed the bug about 60% of the time.

When inspecting the buggy UI, no DOM elements appear to be involved– the elements are simply clipped / hidden behind a background-colored box. Note that triggering repaints from the DOM inspector (toggling e.g. the box-shadow on the modal container) will cause the bug to resolve temporarily (until the next query). This makes me think that there exists some sort of race condition in the layout/paint logic.

The fix seems pretty random and I wish I had a better answer, but I removed the `-webkit-mask` and was consistently unable to reproduce the clipping problem. `-webkit-mask` is not a commonly used property, so it seems plausible that it's buggy. The mask is a nonessential part of `quick-open`, so we can easily do without and see if we continue to see bug reports.

Reviewed By: ssorallen, bolinfest

Differential Revision: D2913580

fb-gh-sync-id: 1bcc924dda856d5ac3a9fba7036f469c372bd7c2
shipit-source-id: 1bcc924dda856d5ac3a9fba7036f469c372bd7c2