Add a generic side bar container UI and service
Summary:
public

Add a side bar UI and service that takes classes that extend
`React.Component` and renders them were the file-tree typically lives.
Consumers can pass a `command` for which the side bar will listen in
order to show an instance of the given component.

The side bar handles serialization of the active view, the side bar's
width, and its visibility.

Reviewed By: matthewwithanm

Differential Revision: D2901941

fb-gh-sync-id: 75f532dc11c5d8823becfbf4b625a7cee0d62b15
shipit-source-id: 75f532dc11c5d8823becfbf4b625a7cee0d62b15