Disable and re-enable Atom's 'tree-view' appropriately
Summary:
public

When Nuclide's 'file-tree' activates, it must replace Atom's
'tree-view'. If 'file-tree' activates during the initial package
activation, listen for `onDidActivateInitialPackages` and disable
'tree-view' afterwards.

When 'file-tree' is deactivated, check if it is disabled and re-enable
Atom's 'tree-view'. This will re-enable 'tree-view' when the 'nuclide'
package is disabled or uninstalled via Atom's UI.

Fixes #354

Reviewed By: zertosh

Differential Revision: D2922366

fb-gh-sync-id: f3d917a06ff509c18ab5b8606de823df17e3b172
shipit-source-id: f3d917a06ff509c18ab5b8606de823df17e3b172