Introduce prefix-based key navigation
Summary:public

Allows navigating the file tree using keys. This is a frequently requested feature, but I know ssorallen has had some reservations. To accommodate everyone, this is a two-diff stack with D2846850 introducing a setting for toggling this behavior.

The match order mirrors the order of items in the file tree, meaning that pressing a letter will prefer subdirectories starting with a given letter over files, unless no subdirectories starting with that letter are present. I think this is still the best UX here.

The keyboard subscription is unfortunately hacky and a bit verbose, since I haven't found a way to attach a generic keydown listener to the filetree (Atom appears to swallow events from non-focused elements). For now, this is a decent v1. The only way around this is adding a hidden `input` element whose focus is managed by the file tree, and capturing keyboard events through the `input` (this is what quick-open does). Because I need to subscribe explicitly to key events, right now it is only possible to select files starting with a-z.

This feature adds a prefix buffer with a timeout. The prefix only grows if actual matches are found, allowing continuous refinement (instead of having to wait for the timeout and starting over). Given a sample file tree
```
A
  Ba
  Bo
    X
  Ce
  Ci
```
the following behavior should occur:

- with `A` selected, typing any letter key does nothing
- with `Ci` selected, typing `c` will select `Ce`, typing `b` will select `Ba`, and typing `bo` will select 'Bo'
- with `Ce` selected, typing `c` does nothing.
- with `Ce` or any of its siblings selected, typing `x` does nothing.
- with both `Ci` and `Ce` selected at the same time, typing `b` selects `Ba` (reducing the multi-selection to single selection).

Reviewed By: matthewwithanm

Differential Revision: D2846839

fb-gh-sync-id: a4185ddef845e354be96512da9522f683e266d32
shipit-source-id: a4185ddef845e354be96512da9522f683e266d32