Fix mini console scrolling bug in DevTools
Summary:public

There appears to be a bug in our version of Chrome Dev Tools where `overflow-y: scroll` is not being set correctly in all cases. It did eventually appear on the element in question after playing around with the console drawers for a while, but it wasn't deterministic.
This CSS hack ensures that the drawer console contents are always scrollable, and prevents long console content from pushing the remaining UI (e.g. the action bar on the sources panel) up beyond the viewport.

Reviewed By: johnislarry

Differential Revision: D2979470

fb-gh-sync-id: 08d3a3ac69f927993f3e7f0c9400ecb2926eec4a
shipit-source-id: 08d3a3ac69f927993f3e7f0c9400ecb2926eec4a