Fix case when array property has 0 children
Summary:public

Previously this code had some incorrect assumptions about the `numchildren` property.  I first noticed this in the $GLOBALS array, which is a global variable.  It was being reported as an empty array, which is wrong.  The reason is that it contained empty arrays, and even a self reference, which caused some NaN issues and some invariants to be hit.  This fixes the issue and adds some nice representation for recursive members.

Reviewed By: jgebhardt

Differential Revision: D2983346

fb-gh-sync-id: e1fc9d983ecd207fc43ab59caedc43f2c954d0bc
shipit-source-id: e1fc9d983ecd207fc43ab59caedc43f2c954d0bc