Remove path/file separation, simplify rendered output.
Summary:public

The `<basename>  <basepath>` pattern makes results hard to read, and stands in the way of performing efficient incremental searching (because there is a lot of visual churn in the output list). We now simply show the entire path and rely on match highlighting to draw attention to the important parts of each path, while keeping the list more visually stable across refinements.

Previously, we wrapped each highlighted character in a separate `span`. Since the list contents change very rapidly, it makes sense to optimize the number of DOM nodes used by calculating the longest possible substrings that are either highlighted or not.

Before:
{F60232119}

After:
{F60232113}

Reviewed By: hansonw

Differential Revision: D2981293

fb-gh-sync-id: 04d8ca712ff74a96f692646f0905633e5a72379e
shipit-source-id: 04d8ca712ff74a96f692646f0905633e5a72379e