Show HHVM toolbar when we are inside a remote Hack project, not a remote Arcanist project
Summary:public

The HHVM toolbar appeared to be broken for OSS Nuclide.  The reason is that we only allow it to be shown for remote Arcanist projects, which doesn't make much sense AFAICT.  This diff changes the criteria to be remote Hack projects, as determined by the presence of a `.hhconfig` file in an ancestor directory of the active file.

Differential Revision: D3001933

fb-gh-sync-id: 395331571f1b2dc2c07c17e5db941f093a371261
shipit-source-id: 395331571f1b2dc2c07c17e5db941f093a371261