Synchronously check next focus target to close WS modal
Summary:Using `setImmediate` introduces an asynchronous call to a React
component's methods. In this case, clicking a button inside the modal
sometimes threw exceptions if the callback passed to `setImmediate`
was called after the component was unmounted.

Use the standard `event.relatedTarget` to test where the focus is
moving so no asynchronous checking of `document.activeElement` is
needed.

public

Reviewed By: shushz

Differential Revision: D3034246

fb-gh-sync-id: 45405862b2af9b976addcaa9f50102d9b8cee402
shipit-source-id: 45405862b2af9b976addcaa9f50102d9b8cee402