Ignore unsupported message types and fix event context
Summary:Creates a "subclass" of node-inspector's Session in order to ignore some commands (and avoid log noise). I added a test to alert us if `_registerNoopCommands` is removed from node-inspector when we upgrade in the future.

There's also a small source modification to correct the context of the `Runtime.executionContextCreated`, detailed in the README.

public

Reviewed By: zertosh

Differential Revision: D3035304

fb-gh-sync-id: 7d9b0ddb01465da8116687dafd74ebbb738a100e
shipit-source-id: 7d9b0ddb01465da8116687dafd74ebbb738a100e