Replace store reference in Bridge with model reference
Summary:public

`DebuggerModel` knows everything: `DebuggerStore`, `Bridge`, `BreakpointStore`, etc...
Until know, `Bridge` knew about `BreakpointStore`, but not `DebuggerStore`. If the `Bridge` wants to talk to the `DebuggerStore`, it needs to go through the `DebuggerModel`, which it did not know about until now.

Injecting dependencies in the constructor is not great. Instead, this injects the entire `DebuggerModel`, which can be used to access its dependents.

Reviewed By: johnislarry

Differential Revision: D3043234

fb-gh-sync-id: 5f2430eed2e981af987c316ddb2444b5764bb553
shipit-source-id: 5f2430eed2e981af987c316ddb2444b5764bb553