Add ability to copy output to clipboard
Summary:This adds the ability to copy from the output window—both the selected text as well as the entire message.

Originally I was going to put these under a single "Copy" command, however that conflicts with the normal OSX behavior (which automatically selects the word you clicked on when there's no current selection).

public

Reviewed By: johnislarry

Differential Revision: D3060452

fb-gh-sync-id: 704513c60a60db964c6d05018b1e591d3bb00f74
shipit-source-id: 704513c60a60db964c6d05018b1e591d3bb00f74