Simplify RN node executor
Summary:This greatly simplifies the RN node executor.

Previously, messages from the RN app would be translated by `ChildManager` into calls of `Child` methods, which would then create its own message format to send to the executor. Responsibilities for handling these messages were divided between all three. After this change, the RN message format is maintained all the way to executor. Not only does this remove unnecessary transformations, it also makes things a lot easier to understand for people familiar with RN already.

In addition, I've moved the script loading into the executor. This results in a clearer division of responsibilities, with RN sending and receiving messages to the executor and Nuclide only being there to facilitate the transfer. With that in mind, I was able to replace `Child` with a function (`executeRnRequests`) that accepts a stream of RN messages and outputs a stream of results. Since this sums up all of Nuclide's responsibilities here, I'd like to remove `ChildManager` entirely (it's now just an OO interface to `executeRnRequests`) but it's currently being used by the `ExecutorServer`, so I'll hold off until we're able to remove that in favor of the new `DebuggerProxyClient`.

This refactor also allowed me to switch from calling `process._debugProcess` on a running process to starting it with `--debug-brk` which fixed the issue with the debugger not skipping the loader breakpoint when using node 5.x (since "restartframe" is no longer called?see https://github.com/nodejs/node/issues/5221).

public

Reviewed By: ssorallen

Differential Revision: D3073832

fb-gh-sync-id: 3604c2d19c5ab336b854766ff10c5a77695bade5
shipit-source-id: 3604c2d19c5ab336b854766ff10c5a77695bade5