Kill debugger on server disconnect
Summary:We currently try to infer how to kill the debugger from directory paths, but after talking with
yinghuitan, I see the right thing to do is kill the debugger based on the related server connection.  So this
diff changes the listener to be at a server-level, not a project-level.

public

Reviewed By: yinghuitan

Differential Revision: D3074488

fb-gh-sync-id: b68443abcdcacee2b3787730191cf8f05254bc02
shipit-source-id: b68443abcdcacee2b3787730191cf8f05254bc02