Implement async Repository API used in v1.7.0-beta0
Summary:public

Atom v1.7.0 implements an async version of `GitRepository` that is
available on sync versions as `#async`. Nuclide's Mercurial
implementation needs to implement the same API in order to be compatible.

Methods that will be superceded by their async counterparts are now
marked as deprecated but will still work for the foreseeable future.

Fixes #439

Reviewed By: mostafaeweda

Differential Revision: D3113130

fb-gh-sync-id: 87b64cf0bb814f28baa5f159162082baa7a231b1
fbshipit-source-id: 87b64cf0bb814f28baa5f159162082baa7a231b1