Remove unnecessary css from outline view
Summary:I swear this had some utility when I added it but I guess with the various
recent changes it no longer does. The `width: 100%` also had the unfortunate
effect of making the highlighting always be exactly the width of the outline
view panel, and it would scroll to the left if you had to scroll sideways to see
the entire outline.

Now, the entire line will be highlighted even when you scroll sideways.

Reviewed By: ssorallen

Differential Revision: D3121168

fb-gh-sync-id: 7eef8e7afd4a5ea969537332c65ab4475c63e2eb
fbshipit-source-id: 7eef8e7afd4a5ea969537332c65ab4475c63e2eb