Move "Nuclide" menu before "Window" and "Help"
Summary:This moves the "Nuclide" menu before "Window"  in the menu bar. I decided to use `add()` and move the item (rather than just manipulating the template directly from the start) because:

1. We don't have to use the undocumented `remove()` to get a corresponding disposable
2. It accounts for an already-existent "Nuclide" menu (just in case)
3. `update()` is debounced by Atom so perf impact of multiple calls shouldn't be an issue.

If there's no menu item with the role "window," the "Nuclide" menu will fall back to being the last item.

@public

Reviewed By: zertosh

Differential Revision: D3115905

fb-gh-sync-id: 8a646bad98c65a1d1a747df6e29eb687122b22a6
fbshipit-source-id: 8a646bad98c65a1d1a747df6e29eb687122b22a6