Use ES module exports when exporting imports
Summary:This change is part of the conversion to ES modules. These are the only modules that `export` an `import`. Switching from `module.exports` is hard to codemod for this. So since there are only a handful of them. I decided to do them manually.

 ---
EDIT:

For the reasons in D3021470#13, we can't use `export {?} from '?';`, so I switched to

```
import {foo} from 'bar';
export {foo};
```

Reviewed By: matthewwithanm

Differential Revision: D3125090

fb-gh-sync-id: 326d25e0e4c0814d232f44ed051ef3c8ce72999f
fbshipit-source-id: 326d25e0e4c0814d232f44ed051ef3c8ce72999f