Use file: scheme for eslint-plugin-nuclide-internal
Summary:`npm shrinkwrap` dies when you try to update a pre-existing shrinkwrap WITH a `file:` dep in `node_modules`. And it works just fine with `./`. However, `apm publish` doesn't like `./` - we figured this out when we tried to published 0.127.0.

The steps I came up for creating pristine shrinkwrap files is also immune to the `./` issue. I didn't know this at the time of D3101998. So now I'm putting it back so `apm publish` works as expected.

Reviewed By: matthewwithanm

Differential Revision: D3128871

fb-gh-sync-id: 71b05cea75edd0d9b4f8e476a04a2e7aa879d486
fbshipit-source-id: 71b05cea75edd0d9b4f8e476a04a2e7aa879d486