Show class fields in flow outline.
Summary:This new outline feature is awesome. But it does not show fields for classes. I think it's a common pattern to define some methods as fields like this:

```js
class Component {
  state = {};

  render() {
    return <button onClick={this.handleClick} />;
  }

  handleClick = (event) => {
  };
}
```

It would be nice if such methods would be shown as well. This PR draft makes above class to appear like this:

```
class Component
  state=
  render()
  handleClick=(event)
```

If you are cool with this change, I'll add tests.
Closes https://github.com/facebook/nuclide/pull/441

Reviewed By: jgebhardt

Differential Revision: D3133571

Pulled By: nmote

fb-gh-sync-id: f7b6204f5ec2b69d20eafec997d0febabc8557ea
fbshipit-source-id: f7b6204f5ec2b69d20eafec997d0febabc8557ea