Diff 7.5: Improve flow types for framework marshalling.
Summary:The flow types for the transformers was incorrect, and relied on flows
erasure of generics to type check. This diff removes the incorrect type annotations
for transformer args/return val and adds correct type parameters for the
MarshallingContext. Note that the MarshallingContext is (currently) very different
for client vs/ server.

Reviewed By: mostafaeweda

Differential Revision: D3135069

fb-gh-sync-id: d63a9bbd1b113363b67a3a5763ae71919ed26b7d
fbshipit-source-id: d63a9bbd1b113363b67a3a5763ae71919ed26b7d