Diff 8: Do not modify remote objects, and hide _interface member.
Summary:This cleans up the interface to the ObjectRegistry, and hides
the _interface member which was being used in the ServerComponent.

Also, stops storing the _interface and _remoteId members on the remote object and moves them to external maps.
This will prevent any potential conflict with user defined members of those names.

Reviewed By: mostafaeweda

Differential Revision: D3129795

fb-gh-sync-id: a5d0737b0cbfe4558ea268f2e52d313e59d1c627
fbshipit-source-id: a5d0737b0cbfe4558ea268f2e52d313e59d1c627