Refactor file-tree
Summary:@public
Complete overhaul of the internal structures backing the file-tree. The new structures are hierarchical and largely immutable.
These data structures allow efficient React rendering of the tree with heavy use of the .shouldComponentUpdate()

As a consequence, the React rendering, for large trees from being a huge bottleneck became computationally insignificant.

The new structures can be efficiently used for more complex functionality, such as on the fly filtering, windowing and so on.

Reviewed By: ssorallen

Differential Revision: D3124632

fb-gh-sync-id: 504c59e8a3c65d715f460b5bb8eb0b6099fd8dde
fbshipit-source-id: 504c59e8a3c65d715f460b5bb8eb0b6099fd8dde