Keep debugger to the left of the tool-bar
Summary:I keep my tool-bar on the right, and both it and the debugger use the default
priority (100). So, it's non-deterministic which one will end up on the left.
Prior to this change, If I opened the debugger panel, then toggle the tool-bar,
it would come back on the left of the debugger which looks really weird.

Reviewed By: johnislarry

Differential Revision: D3146766

fb-gh-sync-id: f258b911e578f174fb1ed253c529899fb538cf6f
fbshipit-source-id: f258b911e578f174fb1ed253c529899fb538cf6f