Fix php debugger crashes
Summary:The problem:

An atom redbox shows up when debugging after doing things like setting or removing breakpoints.

The repro:

On master, have a few tabs in the browser open to your sandbox.  Attach the php debugger and wait for a moment.  Set and remove some breakpoints until the redbox appears.

The fix + context:

As best I can tell this is due to nuclide sending debugger commands to HHVM when the connection is in 'stopping' mode (the mode that comes after code is done executing, but before the connection is killed, usually used for post-mortem features, although we don't use it for anything).  It's a bit hard to tell from the logs, but it seems that our sandbox bombards HHVM with many short lived requests to random endpoints, and it's not too hard to do something like set or remove a breakpoint or any other debugger action when these requests reach their 'stopping' phase.  HHVM apparently doesn't like this so much, and will immediately kill the connection from its end with an ECONNRESET, which we handle somewhat gracefully by displaying a redbox with a generic error message saying the connection was killed.  If this redbox happens it should be possible to keep debugging (only the bad connection was killed) but I don't remember if I tried, and it's possible that this error would just put nuclide into a bad state.

When we enter the 'stopping' state in ConnectionMultiplexer we issue a 'run' command, which by my reading of the xdebug spec, seems wrong.  Instead, I think the correct thing to do is to just send a 'stop' message which will kill the connection.  Being in the 'run' state allows us to send commands to HHVM via nuclide, which is never what we want to do when we are at or after the 'stopping' state.

Reviewed By: yinghuitan

Differential Revision: D3143294

fb-gh-sync-id: 9a06feab6efc4e61f5bcbfc6d3f87faefcb581c2
fbshipit-source-id: 9a06feab6efc4e61f5bcbfc6d3f87faefcb581c2