Introduce nuclide-file-tree.context-menu service.
Summary:This enables us to have better control over the ordering of items in
the context menu for the file tree. While here, I also overhauled the
positioning of the items in the	context menu for the file tree.

The idea behind this new approach is explained in the overview comment
for the definition of the `FileTreeContextMenu` class. You can see this API
in use for real in D2838307.

Note that using this API still feels like a bit of a kludge. I think we can improve
it if we add support for declaring a callback function on `atom$ContextMenuItem`
that can be used in lieu of the `command`. This would eliminate the problem in the
current implementation where an illegitimate command must be added to the registry
(such as `'command-that-should-only-be-fired-from-the-context-menu'` in the
example in the overview comment) simply to satisfy the existing Atom API.
I have not started this conversation with Atom yet, so I don't want to gate this work
on a new API from Atom. On the contrary, I think we'll be able to have a higher
quality discussion with Atom if we can get this code into Nuclide and then call it out
as a pain point that needs to be addressed.

Reviewed By: ssorallen

Differential Revision: D2837364

fb-gh-sync-id: 35cbada72f25ba68f4dc29cedc1e1b74e334e8d6
fbshipit-source-id: 35cbada72f25ba68f4dc29cedc1e1b74e334e8d6