Update RN packager script search and messaging
Summary:The primary purpose of this commit is to update the error messaging to be more helpful. Previously, you could technically satisfy its stated conditions by having RN in your package.json but not installed.

The search now looks for a node_modules directory with RN, and the error reflects that. I broke `getCommandFromNodePackage` into two functions for clarity's sake.

Reviewed By: ssorallen

Differential Revision: D3156557

fb-gh-sync-id: ffd71ba72a6013920fe5dd7850150003b5ae61c7
fbshipit-source-id: ffd71ba72a6013920fe5dd7850150003b5ae61c7