Fix a bug where atom-script did not successfully flush everything written to console.log().
Summary:I'll be honest: I am not 100% thrilled with this solution. I believe there is still a race
condition, as noted in `nuclide-atom-script/bin/main.js`. However, this gets us to a much
better place than we were in before. Overall, `atom-script` has a hacky implementation to work around
the issue that it's not possible to write clean output to stdout from Atom using `console.log()`.
The ideal way to eliminate the hacks is to work with Atom to get the functionality we need.
This will have to do for now.

Originally, the markdown example for `atom-script` required an `--out` parameter
because writing to stdout was not an option. Specifically, when I tried using
`console.log()` to write the output, the `atom` process would get killed before
all of the data was flushed.

Now, I can run the following and all of the generated HTML is written to stdout, as expected:

```
~/.atom/packages/nuclide/pkg/nuclide-atom-script/bin/atom-script \
  ~/.atom/packages/nuclide/pkg/nuclide-atom-script/samples/markdown.js \
  ~/.atom/packages/nuclide/pkg/nuclide-atom-script/README.md
```

This revision also stops forwarding Atom's internal messages to stderr because it's pretty distracting.

Reviewed By: zertosh

Differential Revision: D3158718

fb-gh-sync-id: 549bb800d44e9eb904b3d77ac6333861412a863a
fbshipit-source-id: 549bb800d44e9eb904b3d77ac6333861412a863a