Ensure the debugger listens to the correct ports
Summary:D3084171 Introduced a hack around an issue, and this diff removes that hack and implements the correct solution.

The issue: we changed the hhvm config on people's devservers so that when they launched php scripts, these scripts sent xdebug requests to a different port than the port we use for attaching to hhvm.  This was fine, except that it broke the dummy connection when we would attach, because the dummy connection would try to use the port we specified in the config file I changed, instead of the port used for attaching.

The fix: Check if we are in a launch or attach scenario.  If we are launching, spin up one dbgpConnector to listen to the script launching port.  This works for both dummy connection and launched script connection.  If we are attaching, then spin up two DbgpConnectors.  One to listen to the dummy connection's hhvm and one to listen to the webserver hhvm for attaching.  The rest of the code is just managing these object's lifetimes and other state, and cleaning up the previous hack.

Also the dummy connection is ready much faster now, due to the speedup from switching config files.

Reviewed By: yinghuitan

Differential Revision: D3088095

fb-gh-sync-id: c553dbefc50971773ad17509e6c06070522ba84d
fbshipit-source-id: c553dbefc50971773ad17509e6c06070522ba84d