Diff 15: Add conection shutdown.
Summary:Finally, change the shutdown message to a close connection message with
shutdown parameter. When closing a connection, always dispose any orphaned RPC
objects which have been marshalled back to the client.

Also changes the 'this' argument of service framework methods to be the SocketClient
which is per-connection rather than the ServerComponent which is per-server.

Reviewed By: mostafaeweda

Differential Revision: D3136718

fb-gh-sync-id: 8f417bf1b69bd0200bc21908dd042282e5e2b3de
fbshipit-source-id: 8f417bf1b69bd0200bc21908dd042282e5e2b3de